+++ src/Adapter/MSTestAdapter.PlatformServices/Services/TestContextImplementation.cs
@@ -44,6 +44,14 @@ public sealed class TestContextImplementation : TestContext, ITestContext, IDisp
             => _builder.Append(buffer, index, count);
 
         [MethodImpl(MethodImplOptions.Synchronized)]
+        internal void AppendLine(string? value)
+            => _builder.AppendLine(value);
+
+        [MethodImpl(MethodImplOptions.Synchronized)]
+        internal void Clear()
+            => _builder.Clear();
+
+        [MethodImpl(MethodImplOptions.Synchronized)]
         public override string ToString()
             => _builder.ToString();
     }
@@ -70,7 +78,7 @@ public sealed class TestContextImplementation : TestContext, ITestContext, IDisp
     private SynchronizedStringBuilder? _stdOutStringBuilder;
     private SynchronizedStringBuilder? _stdErrStringBuilder;
     private SynchronizedStringBuilder? _traceStringBuilder;
-    private StringBuilder? _testContextMessageStringBuilder;
+    private SynchronizedStringBuilder? _testContextMessageStringBuilder;
 
     /// <summary>
     /// Unit test outcome.
@@ -421,7 +429,7 @@ public sealed class TestContextImplementation : TestContext, ITestContext, IDisp
         return _traceStringBuilder;
     }
 
-    private StringBuilder GetTestContextMessagesStringBuilder()
+    private SynchronizedStringBuilder GetTestContextMessagesStringBuilder()
     {
         // Prefer writing to the current test context instead of 'this'.
         // This is just a hack to preserve backward compatibility.
@@ -429,7 +437,7 @@ public sealed class TestContextImplementation : TestContext, ITestContext, IDisp
         // Then, they write to the "wrong" test context.
         // Here, we are correcting user's fault by finding out the correct TestContext that should receive the message.
         TestContextImplementation @this = CurrentTestContext ?? this;
-        _ = @this._testContextMessageStringBuilder ?? Interlocked.CompareExchange(ref @this._testContextMessageStringBuilder, new StringBuilder(), null)!;
+        _ = @this._testContextMessageStringBuilder ?? Interlocked.CompareExchange(ref @this._testContextMessageStringBuilder, new SynchronizedStringBuilder(), null)!;
         return @this._testContextMessageStringBuilder;
     }