+++ src/Analyzers/MSTest.Analyzers/Helpers/WellKnownTypeNames.cs
@@ -52,7 +52,9 @@ internal static class WellKnownTypeNames
     public const string SystemIAsyncDisposable = "System.IAsyncDisposable";
     public const string SystemIDisposable = "System.IDisposable";
     public const string SystemReflectionMethodInfo = "System.Reflection.MethodInfo";
+    public const string SystemString = "System.String";
     public const string SystemThreadingCancellationToken = "System.Threading.CancellationToken";
+    public const string SystemThreadingCancellationTokenSource = "System.Threading.CancellationTokenSource";
     public const string SystemThreadingTasksTask = "System.Threading.Tasks.Task";
     public const string SystemThreadingTasksTask1 = "System.Threading.Tasks.Task`1";
     public const string SystemThreadingTasksValueTask = "System.Threading.Tasks.ValueTask";