+++ src/Platform/Microsoft.Testing.Platform/OutputDevice/Terminal/TerminalTestReporter.cs
@@ -62,7 +62,7 @@ internal sealed partial class TerminalTestReporter : IDisposable
     private bool WasCancelled
     {
         get => field || _testApplicationCancellationTokenSource.CancellationToken.IsCancellationRequested;
-        set => field = value;
+        set;
     }
 
     private bool? _shouldShowPassedTests;
@@ -159,6 +159,11 @@ internal sealed partial class TerminalTestReporter : IDisposable
         _terminalWithProgress.WriteToTerminal(_isDiscovery ? AppendTestDiscoverySummary : AppendTestRunSummary);
 
         NativeMethods.RestoreConsoleMode(_originalConsoleMode);
+
+        // This is relevant for HotReload scenarios. We want the next test sessions to start
+        // on a new TestProgressState
+        _testProgressState = null;
+
         _buildErrorsCount = 0;
         _testExecutionStartTime = null;
         _testExecutionEndTime = null;